+++ feature/tengwar/domain/lib/get-encoding.ts
