+++ main.ts
@@ -1,8 +1,8 @@
-import { ProcessTengwar } from 'core/types';
+import { TengProcessor } from 'core/types';
 import { DEFAULT_PLUGIN_SETTINGS } from 'feature/settings/domain/entity/default-plugin-settings';
 import { PluginSettings } from 'feature/settings/domain/entity/plugin-settings';
 import { SettingsTab } from 'feature/settings/ui/settings-tab';
-import { processTengwar } from 'feature/tengwar/ui/teng-processor';
+import { tengProcessor } from 'feature/tengwar/ui/teng-processor';
 import { Plugin } from 'obsidian';
 
 export default class TengwarObsidianPlugin extends Plugin {
@@ -18,7 +18,7 @@ export default class TengwarObsidianPlugin extends Plugin {
       element.className = '';
       element.id = '';
 
-      processTengwar(source, element as HTMLElement, this.settings);
+      tengProcessor(source, element as HTMLElement, this.settings);
     });
   }
 
@@ -27,14 +27,14 @@ export default class TengwarObsidianPlugin extends Plugin {
     /**
      * Add code block processor for 'teng'
      */
-    this.reg('teng', processTengwar);
+    this.reg('teng', tengProcessor);
     /**
      * This adds a settings tab so the user can configure various aspects of the plugin
      */
     this.addSettingTab(new SettingsTab(this.app, this));
   }
 
-  reg(language: string, processor: ProcessTengwar) {
+  reg(language: string, processor: TengProcessor) {
     this.registerMarkdownCodeBlockProcessor(language, (source, el) => {
       processor(source, el, this.settings);
     });
